add status entity and more fields to booking

+++ src/main/resources/db/migration/V202301131501__create_users_roles_table.sql
@@ -3,7 +3,7 @@ DROP TABLE IF EXISTS public.users_roles;
 CREATE TABLE users_roles
 (
     user_id BIGINT REFERENCES users (id),
-    role_id BIGINT REFERENCES roles (id),
+    role_id INTEGER REFERENCES roles (id),
     PRIMARY KEY (user_id, role_id)
 );
